Admins can post media reviews by pasting a link
Admins can now post reviews by pasting a link instead of copying the
review title, author, and article text. Given the URL, we use the
Alchemy API to extract the title and author and run sentiment
analysis on the review text.

After creating a review, we display it so the user can verify that the
API correctly parsed the title, author, and text.

Also, since Alchemy's sentiment analysis returns a single value, we've
removed the two unused sentiment columns from `media_reviews`.

**Setup Alchemy API**
* Add Alchemy Gem to `Gemfile`
* Add Alchemy API key to `.env.sample`
* Create initializer for Alchemy config

**Refactor application to use Alchemy for metadata and analysis**
* Create `MediaReviewMetadataExtractor` to extract title and author and perform sentiment analysis using Alchemy
* Remove unused `SentimentAnalyzer` and columns from `media_reviews`
* Refactor `MediaReview` for use with new `MediaReviewMetadataExtractor` and single sentiment column
* Refactor `Admin::MediaReveiwsController` to use metadata extractor

**View changes**
* Remove unused sentiment fields from `media_review` partial
* Create `admin/media_reviews#create` view to display the saved review
* Remove unused `title`, `author`, and `text` fields from `admin/media_reviews#new`
* Retain fields on `admin/media_reviews#edit` form to allow users to make corrections

https://trello.com/c/o9o7dA63