Add feature specs for current events dashboard
Additionally, refactor `factories.rb` to make events with user and
media reviews traits of the events factory, instead of sub-factories.
This allows easier creation of events with both types of reviews.

Note: The helper methods in this spec are intended to be reusable across
other dashboard specs, and can be extracted when they are written.