Infer Media Review Source from URL
Admins no longer have to manually type the source when creating a
Media review; the source is inferred based on the URL.

This commit adds a `UrlSourceExtractor` class to determine the source
name from a given URL, and updates the `MediaReviewMetadataExtractor`
class to extract the source in addition to other metadata.

Additionally, this commit adds unit specs for
`MediaReviewMetadataExtractor` and refactors.

https://trello.com/c/80R2da0x